refactored User Account Settings
